requirements.txt
# Has to be early, as others depend on it
wheel

# For SQLite support
apsw

# Date parsing
python-dateutil

# Configuration file
PyYAML

# Web requests
requests

###########################################################################
# Specific output of 'pip-chill'
#
# This serves as a reference of all the versions of python modules that
# this was last tested under.
###########################################################################
# aiohttp==1.2.0
# appdirs==1.4.0
# apsw==3.13.0.post1
# arrow==0.10.0
# async-timeout==1.1.0
# beautifulsoup4==4.5.3
# checkdmarc==2.7.0
# decorator==4.0.11
# devscripts==2.17.6+deb9u2
# feedgen==0.7.0
# feedparser==5.1.3
# grip==4.5.2
# html5lib==0.999999999
# httplib2==0.9.2
# imgurpython==1.1.7
# iotop==0.6
# jinjaapidoc==0.4.0
# jwt==0.6.1
# keyring==10.1
# keyrings.alt==1.3
# limnoria==2017.8.18
# multidict==2.1.4
# numpy==1.12.1
# packaging==16.8
# parsedmarc==6.8.2
# pip-chill==1.0.0
# psycopg2==2.6.2
# pyaml==16.12.2
# pyasn1==0.1.9
# pycrypto==2.6.1
# pycurl==7.43.0
# pygobject==3.22.0
# pyopenssl==16.2.0
# pyparsing==2.1.10
# pyrss2gen==1.1
# python-apt==1.4.1
# python-debian==0.1.30
# python-debianbts==2.6.1
# python-gnupg==0.3.9
# pytwitcherapi==0.9.3
# pyyaml==3.12
# reportbug==7.1.7
# roman==2.0.0
# sphinx-rtd-theme==0.1.9
# sqlalchemy==1.0.15
# timeout-decorator==0.4.0
# unidecode==0.4.19
# virtualenv==15.1.0
# webencodings==0.5
# yarl==0.8.1
###########################################################################
